requirements.txt
streamlit
google-cloud-firestore
firebase_admin
openpyxl